aurora/requirements.txt
numpy>=1.23.4,<2
pandas>=1.5.1,<2
Pillow>=9.3.0,<10
Shapely>=2,<3
rasterio>=1.3.3,<2
marshmallow_dataclass>=8.5.9,<9
tqdm>=4.64.1,<5
imagehash>=4.3.1,<5
scikit-learn>=1.1.3,<2
typeguard>=2.13.3,<3
pycocotools>=2.0.6,<3